Making the OkHttp internal framed tests more stable
The Spdy3ConnectionTest experienced some flakyness and inspecting
it things could be improved.

First of all the MockSpdyPeer.play() function was returning
before the peer task executes, allowing room for race conditions.
This has been fixed with the use of a CountDownLatch, so play()
now waits for the task to start handling HTTP requests.

Another potential problem is that tests would share the
instance of MockSpdyPeer and rely on tearDown to clear any data,
or resources. Tests have been updated to have their own instance
of the class which gets auto-closed on test end.

A last issue found is one where a listener would first provide an
HTTP response before marking that the callback was received. This
could cause a race condition as the main thread may start
handling the response and checking if the callback was indeed
received. It got fixed by first marking the callback and then
providing the response.

Bug: 191097565
Test: atest CtsLibcoreOkHttpTestCases --iterations 300
Test: atest CtsLibcoreOkHttpTestCases:com.squareup.okhttp.internal.framed.Spdy3ConnectionTest --iterations 300
Test: atest CtsLibcoreOkHttpTestCases:com.squareup.okhttp.internal.framed.Http2ConnectionTest --iterations 300

Change-Id: I88954f470d67299ff3c865a951e1ad91c7c8d1dd