Allow okhttp-norepackage to compile in master-art
Previously, it failed to build in master-art because it could not find
the "android_stubs_current" module which is defined in frameworks/base
and so not available in master-art.

This switches the sdk_version to "core_current" which causes the build
to fail when building the target variant because it cannot find the
"android.util.Log" class. That was already an issue for host and was
fixed by building host against a placeholder stubs class provided by
"okhttp-android-util-log". Making the target variant build against that
as well fixed the build.

This is needed because of a change to the ART sdk to include the target
variant of okhttp-norepackage caused it to be built, whereas previously
it was not built.

Bug: 193044041
Test: m okhttp-norepackage
Change-Id: Idd0acb7e2221f720f1d28d5b767009b41b76e881