Fixing stability of headersOnlyStreamIsClosedAfterReplyHeaders
Spdy3ConnectionTest.headersOnlyStreamIsClosedAfterReplyHeaders
was occasionally failing. The error seems to be due to the fact
that when a new stream is created, if only headers are exchanged
and the exchange happens too quickly, then the stream is not
marked as open, which causes the test to fail, as it expects an
open stream after creation.

This was fixed in upstream by removing the check that fails due
to race conditions.

See https://github.com/square/okhttp/commit/49dfda04ae9df1b915ccea17ca1e70ff48e06e6c

Bug: 191097565
Test: atest
CtsLibcoreOkHttpTestCases: com.squareup.okhttp.internal.framed.Spdy3ConnectionTest --iterations 1000
Change-Id: Ic73d98407c617ea86bf4a6689109d55240043670