[DO NOT MERGE] Update external/okhttp/repackaged to match source.
The fix for CVE-2021-0341 in qt-dev was cherry-picked from
pi-dev. However while the repackaged okhttp files did not
required merging for pi-dev, they *do* for qt-dev and beyond. This
change merges them in qt-dev, and qt-qpr2-dev after automerging.

Is set to do not merge as the fix was applied correctly to
rvc-dev and all branches downstream of that.

Bug: 171980069
Test: atest CtsLibcoreTestCases CtsLibcoreOkHttpTestCases
Change-Id: Ic47bb42d68c768277be0a8677cd9830ee03ff293